requirements.txt
arrow==1.2.3
binaryornot==0.4.4
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
idna==3.4
inflection==0.5.1
Jinja2==3.1.2
jinja2-time==0.2.0
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
py-ulid==1.0.3
pydantic==1.10.8
Pygments==2.15.1
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0
requests==2.31.0
rich==13.4.1
shellingham==1.5.1
six==1.16.0
text-unidecode==1.3
typer==0.9.0
typing-extensions==4.6.3
urllib3==2.0.2
